add notification model

+++ app/assets/javascripts/controllers/areas.js
@@ -6,38 +6,70 @@
       function($scope, areas, Area,ModalService) {
           $scope.areas = areas;
           console.log($scope.areas);
+
           $scope.addNewArea = function () {
               ModalService.showModal({
-                  templateUrl: "areas/new-area-modal/new-area-modal.html",
-                  controller: "newAreaModalCtrl",
+                  templateUrl: "areas/area-modal/area-modal.html",
+                  controller: "areaModalCtrl",
                   inputs:{
-                    areas:areas
+                    title: 'Add new area',
+                    area: {}
                   }
               }).then(function(modal) {
-                  // The modal object has the element built, if this is a bootstrap modal
-                  // you can call 'modal' to show it, if it's a custom modal just show or hide
-                  // it as you need to.
                   modal.element.modal();
-                  modal.close.then(function(result) {
-                      $scope.message = result ? "You said Yes" : "You said No";
+                  modal.close.then(function(area) {
+                      if(!_.isEmpty(area)){
+                          new Area(area).create().then(function (response) {
+                              console.log(response);
+                              areas.push(response);
+                          },function (response) {
+                              console.log(response);
+                          });
+                      }
                   });
               });
           };
 
           $scope.destroyArea = function (area) {
               var defaultArea = Area.defaultArea($scope.areas);
-              _.forEach(area.devices, function (device) {
-                  device.areaId = defaultArea.id;
-                  device.update();
-              });
-              defaultArea.devices = _.concat(defaultArea.devices,area.devices );
-              area.delete().then(function (response) {
-                  _.pull($scope.areas, area);
-                  console.log(response);
-              },function (response) {
-                  console.log(response);
-              })
+              if (area.id != defaultArea.id) {
+                  _.forEach(area.devices, function (device) {
+                      device.areaId = defaultArea.id;
+                      device.update();
+                  });
+                  defaultArea.devices = _.concat(defaultArea.devices, area.devices);
+                  area.delete().then(function (response) {
+                      _.pull($scope.areas, area);
+                      console.log(response);
+                  }, function (response) {
+                      console.log(response);
+                  })
+              }
           };
+
+          $scope.editArea = function (area) {
+              ModalService.showModal({
+                  templateUrl: "areas/area-modal/area-modal.html",
+                  controller: "areaModalCtrl",
+                  inputs:{
+                      title: 'Edit area',
+                      area:area
+                  }
+              }).then(function(modal) {
+                  modal.element.modal();
+                  modal.close.then(function(updatedArea) {
+                      if(!_.isEmpty(updatedArea)){
+                            updatedArea.update().then(function (response) {
+                                _.assign(area, updatedArea);
+
+                                console.log(response);
+                          },function (response) {
+                              console.log(response);
+                          });
+                      }
+                  });
+              });
+          }
       }
     ]);
